Refactored redux state into types folder
