+++ .env
@@ -1,4 +1,3 @@
 PORT=4000
 REACT_APP_GOOGLE_CLIENT_ID=""
 REACT_APP_DEV_TOKEN="TEST"
-REACT_APP_AC_API_URI="http://localhost:3000/dev"
