+++ src/components/NotebookCell.tsx
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import { StyleSheet, css } from 'aphrodite';
 import { Icon } from 'rsuite';
 
-import { ReduxState } from '../redux';
+import { ReduxState } from '../types/redux';
 import { _editor } from '../redux/actions';
 import { EditorCell, ImmutableEditorCell } from '../types/notebook';
 import { palette, spacing } from '../constants/theme';
