+++ src/pages/EditorPage/LeftSidebar/KernelPanel.tsx
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import { StyleSheet, css } from 'aphrodite';
 import { Button, Icon, IconButton, Input, Modal, Popover, Whisper } from 'rsuite';
 
-import { ReduxState } from '../../../redux';
+import { ReduxState } from '../../../types/redux';
 import { _editor } from '../../../redux/actions';
 import { palette, spacing, timing } from '../../../constants/theme';
 import { DEFAULT_GATEWAY_URI } from '../../../constants/jupyter';
