+++ src/pages/EditorPage/RightSidebar/DownloadsPanel.tsx
@@ -4,7 +4,7 @@ import { StyleSheet, css } from 'aphrodite';
 import { Button, Icon } from 'rsuite';
 
 import { spacing } from '../../../constants/theme';
-import { ReduxState } from '../../../redux';
+import { ReduxState } from '../../../types/redux';
 import { download } from '../../../utils/notebook';
 
 const styles = StyleSheet.create({});
