+++ src/pages/EditorPage/hooks/useNotebooks.ts
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
-import { ReduxState } from '../../../redux';
+import { ReduxState } from '../../../types/redux';
 import { _editor } from '../../../redux/actions';
 import { isOlderThan } from '../../../utils/date';
